test: :white_check_mark: split server.ts and test
